Map Generation++
Fixed "symmetrical" maps. Problem was in the DLL. Now the growth is
truly random.

Next, I want to implement a "spawnRiver" function that will pick a
random edge (x is either 0 or MAP_DIMENSION) or (y is either 0 or
MAP_DIMENSION). From there, it will follow a psuedo random riverlike
trail till it hits one of the edges again. During its travel it
thickens, but the grow_seed function should be used after it as well.